Adds certificate expiration days functionality
Allows specifying the certificate expiration duration.

This change introduces the ability to define the validity period
for generated certificates by adding a `CertExpireDays` property
to the request model and passing it to the EasyRSA service.
A default value of 365 days is set and used when no value provided.
Also improves certificate handling by checking if a request file exists
before generating a new certificate. If the file exists the existing request is re-signed.
CRL update ensures that revoked certificates are properly handled.